- updated doc for FT_New_Memory_Face - removed lots of compiler warnings in lint-style   warning modes (/W4 with Visual C++)

+++ src/psaux/t1decode.c
@@ -284,7 +284,7 @@
       FT_Outline  dummy;
 
 
-      dummy.n_points = base->n_points - n_base_points;
+      dummy.n_points = (short)(base->n_points - n_base_points);
       dummy.points   = base->points   + n_base_points;
 
       FT_Outline_Translate( &dummy, adx - asb, ady );
